requirements.txt
cryptography==38.0.1
PyJWT==2.4.0
python-dotenv==0.21.0
requests==2.28.1
Flask==2.2.3
pyairtable==1.4.0
gunicorn==19.10.0
pytz==2021.3